Use correctly sized variable to fix warnings

+++ src/zip.h
@@ -436,7 +436,7 @@ typedef struct
     ZPOS64_T totalUncompressedData;
 #ifndef NOCRYPT
     unsigned long keys[3];     /* keys defining the pseudo-random sequence */
-    const unsigned long* pcrc_32_tab;
+    const uint32_t* pcrc_32_tab;
     int crypt_header_size;
 #endif
 } curfile64_info;
